+++ crates/goose-cli/src/session/commands/mod.rs
@@ -1,6 +1,8 @@
 pub mod context;
 pub mod extension;
+pub mod mode;
 pub mod plan;
 pub mod prompt;
 pub mod recipe;
 pub mod settings;
+pub mod ui;
